+++ src/pages/categoryCombos/New.tsx
@@ -8,15 +8,32 @@ import { initialValues } from './form/categoryComboSchema'
 
 const section = SECTIONS_MAP.categoryCombo
 
-export const Component = () => {
+export const Component = ({
+    onSubmitted,
+    footer,
+    redirectOnSubmitted = true,
+}: {
+    readonly onSubmitted?: () => void
+    readonly footer?: React.ReactNode
+    readonly redirectOnSubmitted?: boolean
+} = {}) => {
     return (
         <FormBase
-            onSubmit={useOnSubmitNew({ section })}
+            onSubmit={useOnSubmitNew({
+                section,
+                onSubmitted,
+                redirectOnSubmitted,
+            })}
             initialValues={initialValues}
             validate={validate}
             includeAttributes={false}
+            section={section}
         >
-            <DefaultNewFormContents section={section}>
+            <DefaultNewFormContents
+                section={section}
+                footer={footer}
+                dataTest="categoryComboNewForm"
+            >
                 <CategoryComboFormFields />
             </DefaultNewFormContents>
         </FormBase>