+++ src/pages/dataElements/merge/index.ts
@@ -1,0 +1,2 @@
+export * from './DataElementMergeFormFields'
+export * from './DataElementMergeSchema'