+++ src/pages/indicatorTypes/Merge.spec.tsx
@@ -2,11 +2,13 @@ import { render } from '@testing-library/react'
 import React from 'react'
 import { SECTIONS_MAP } from '../../lib'
 import { testIndicatorType } from '../../testUtils/builders'
-import TestComponentWithRouter from '../../testUtils/TestComponentWithRouter'
+import TestComponentWithRouter, {
+    CustomData,
+} from '../../testUtils/TestComponentWithRouter'
 import { generateDefaultMergeTests } from '../defaultMergeTests'
 import { Component as Merge } from './Merge'
 
-const renderMerge = async () => {
+const renderMerge = async (customData: CustomData = {}) => {
     const routeOptions = {
         handle: { section: SECTIONS_MAP.indicatorType },
     }
@@ -15,7 +17,11 @@ const renderMerge = async () => {
         <TestComponentWithRouter
             path="/indicators/merge"
             customData={{
-                indicatorTypes: [testIndicatorType(), testIndicatorType()],
+                indicatorTypes: {
+                    indicatorTypes: [testIndicatorType(), testIndicatorType()],
+                    pager: { page: 1, pageCount: 1, total: 2, pageSize: 10 },
+                },
+                ...customData,
             }}
             routeOptions={routeOptions}
         >
@@ -23,7 +29,11 @@ const renderMerge = async () => {
         </TestComponentWithRouter>
     )
 }
-generateDefaultMergeTests({ componentName: 'Indicator type', renderMerge })
+generateDefaultMergeTests({
+    componentName: 'Indicator type',
+    mergeResource: 'indicatorTypes/merge',
+    renderMerge,
+})
 
 describe('Indicator type additional tests', () => {
     it('should show a warning if target and source have different factors', () => {})