+++ src/pages/predictors/Form.spec.tsx
@@ -1,0 +1,3 @@
+describe('Common', () => {
+    it.skip('should open the dataElement new form in a drawer when clicking the "Add new" button', () => {})
+})